Create Code
In a productive environment, sorting is a complecated process.Continuous manual sorting create waste of time and manpower. So  we are going to built an color based Sorting Machine using NodeMCU-ESP8266, TCS3200 Colour Sensor, and two Servo Motors, as the title says it will sort the things according to their colour.The TCS3200 Colour Sensor is responsible for sensing the colour of an object and two servo motors are used to put them into the respective predefined colour box.